Cast open and close levels to float

+++ config/EURUSD_M15.h
@@ -20,10 +20,10 @@ struct Stg_Force_Params_M15 : StgParams {
     lot_size = 0;
     signal_open_method = 0;
     signal_open_filter = 2;
-    signal_open_level = 3;
+    signal_open_level = (float)3;
     signal_open_boost = 1;
     signal_close_method = 0;
-    signal_close_level = 0;
+    signal_close_level = (float)0;
     price_stop_method = 0;
     price_stop_level = 0;
     tick_filter_method = 1;
